+++ template/trout/features/step_definitions/js_steps.rb
@@ -1,3 +1,0 @@
-Then 'the javascript expression "$expression" should return "$result"' do |expression, result|
-  page.evaluate_script(expression).to_s.should == result
-end